feat!: changing the direction of theme
removed the warrior and rogue theme and added autumn and spring themes.
previosly they were attempts to recreate other themes with the
highlighting philosophy of makurai, but to no avail.

spring and autumn aim to bring new life to the theme with original
colors schemes that are different then the main